refactor(pages): import from modules
